+++ vulnfeeds/cmd/nvd-cve-osv/main_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/google/osv/vulnfeeds/cves"
-	"github.com/google/osv/vulnfeeds/testutils"
+	"github.com/google/osv/vulnfeeds/internal/testutils"
 )
 
 func TestReposFromReferences(t *testing.T) {
